FS: factor auth types and funcs to a separate file
