+++ packages/file-storage-common/src/index.ts
@@ -9,58 +9,18 @@ import {
 import {
   AttachmentApi,
   AttachmentDirectUploadUrl,
-  Configuration,
-  ConfigurationParameters,
   DirectAttachment,
   DirectAttachmentsApi,
   ThumbnailApi,
 } from '@meeco/vault-api-sdk';
 import axios from 'axios';
 import { AzureBlockUpload } from './azure-block-upload';
+import { IFileStorageAuthConfiguration, buildApiConfig, getHeaders } from './auth';
+
 export { AzureBlockDownload } from './azure-block-download';
 export { AzureBlockUpload } from './azure-block-upload';
 export { BlobStorage } from './services/Azure';
-
-/**
- * @param vault_access_token Meeco Vault Bearer Token.
- * @param delegation_id Original User id of User acting as a delegate.
- * @param subscription_key Meeco API Subscription key.
- */
-export interface IFileStorageAuthConfiguration {
-  vault_access_token?: string;
-  delegation_id?: string;
-  subscription_key?: string;
-  oidc_token?: string;
-}
-
-export function buildApiConfig(
-  auth: IFileStorageAuthConfiguration,
-  vaultUrl: string,
-  fetchApi?: any
-): Configuration {
-  const headers = getHeaders(auth);
-
-  const configParams: ConfigurationParameters = {
-    basePath: vaultUrl,
-    headers,
-  };
-  if (fetchApi) {
-    configParams['fetchApi'] = fetchApi;
-  }
-  return new Configuration(configParams);
-}
-
-/**
- * Extract HTTP headers from an `IFileStorageAuthConfiguration` object.
- */
-function getHeaders(auth: IFileStorageAuthConfiguration): { [header: string]: string } {
-  const headers = {};
-  headers['Meeco-Delegation-Id'] = auth.delegation_id || '';
-  headers['Meeco-Subscription-Key'] = auth.subscription_key || '';
-  headers['Authorization'] = auth.vault_access_token || '';
-  headers['authorizationoidc2'] = auth.oidc_token ? 'Bearer ' + auth.oidc_token : '';
-  return headers;
-}
+export { IFileStorageAuthConfiguration, buildApiConfig };
 
 /**
  *
